Refactor medication retrieval and deletion methods
Updated the method for retrieving and deleting medications to use FindAsync for better performance and added null checks for improved error handling.

+++ challengeFiap/Controllers/MedicamentoesController.cs
@@ -63,7 +63,13 @@ public class MedicamentoesController : ControllerBase
 
         try
         {
-            var medicamento = await _context.Medicamentos.FirstOrDefaultAsync(m => m.Id_medicamento == id_medicamento);
+            var medicamento = await _context.Medicamentos.FindAsync(id_medicamento);
+
+            if (medicamento == null)
+            {
+                _logger.LogWarning("Medicamento não encontrada para o ID");
+                return NotFound("Id de medicamento não encontrado");
+            }
 
             _logger.LogInformation("Medicamento encontrado com sucesso para o ID {IdMedicamento}", id_medicamento);
             return Ok(medicamento);
@@ -193,7 +199,14 @@ public class MedicamentoesController : ControllerBase
 
         try
         {
-            var medicamento = await _context.Medicamentos.FirstOrDefaultAsync(e => e.Id_medicamento == id_medicamento);
+            var medicamento = await _context.Medicamentos.FindAsync(id_medicamento);
+
+            if (medicamento == null)
+            {
+                _logger.LogWarning("Medicamento não encontrada para exclusão.");
+                return NotFound("Id medicamento não encontrado");
+            }
+            
             _context.Medicamentos.Remove(medicamento);
             await _context.SaveChangesAsync();
